fix: - project structure - cyclic serializing problem add: - new program status (+migration) - new methods to repositories - new check for starting program logic - project design description to README.md - handling of sql exceptions etc

+++ src/main/java/ru/lonelydutchhound/remotedevicecontrol/dto/WashingProgramDTO.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 public class WashingProgramDTO implements DTO{
     UUID id;
     String name;
-    byte temperature;
+    int temperature;
     Long duration;
     int spinSpeed;
 }
